% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_set.R
\name{get_index_set}
\alias{get_index_set}
\title{Generate Coalition Labels for a Cooperative Game}
\usage{
get_index_set(n)
}
\arguments{
\item{n}{Integer. The number of players in the game.}
}
\value{
A character vector with all non-empty coalitions represented as strings.
The coalitions are ordered first by size and then lexicographically.
}
\description{
This function generates character labels representing all possible non-empty coalitions
of players in a cooperative game with \code{n} players. Each coalition is represented as a
concatenated string of player indices (e.g., "12", "13", "123").
}
\examples{
get_index_set(3)
# Returns: "1" "2" "3" "12" "13" "23" "123"

}
