% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapley_gamma.R
\name{get_shapley_gamma}
\alias{get_shapley_gamma}
\title{Compute Gamma Weights for the Shapley-Based Metric}
\usage{
get_shapley_gamma(n, last_val = 1)
}
\arguments{
\item{n}{Integer. The number of players in the cooperative game.}

\item{last_val}{Numeric (default = 1). A custom value for the gamma weight of the grand coalition.}
}
\value{
A numeric vector of length \eqn{2^n - 1} with weights \eqn{\gamma_S} for each coalition.
}
\description{
This function returns the vector of weights \eqn{\gamma_S} associated with a special
metric used in cooperative game theory. This metric is minimized by the Shapley value
and is defined as: \deqn{\sqrt{\sum_S \gamma_S (v(S) - x(S))^2}}
where the sum runs over all coalitions \eqn{S}.
}
\details{
These weights are derived from the characterization of the Shapley value as the unique
allocation that minimizes a weighted Euclidean distance from the game vector.

The value of \eqn{\gamma_S} depends on the cardinality \eqn{|S|} of each coalition, and is given by:

\deqn{\gamma_S = \frac{1}{\binom{n-2}{|S|-1}}}

The weight for the grand coalition (the last entry) is set to \code{last_val}.
}
\examples{
get_shapley_gamma(3)

}
