% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_label_game.R
\name{set_label_game}
\alias{set_label_game}
\title{Assign Coalition Labels to a Game Vector}
\usage{
set_label_game(v)
}
\arguments{
\item{v}{A numeric vector representing the values of the coalitions in a cooperative game.
The length must be equal to 2^n - 1 for some integer n (number of players).}
}
\value{
The same vector \code{v}, but with names (labels) assigned to each element.
}
\description{
This function assigns string labels to the elements of a numeric vector \code{v},
which represents the value of a cooperative game. Each label corresponds to
a non-empty coalition of players. The function assumes the vector is ordered
according to the standard ordering of coalitions used in cooperative game theory.
}
\details{
This function is compatible with vectors generated by the CoopGame package.
It does not depend on CoopGame internally, so it can be used independently.
}
\examples{
# For internal testing purposes only:
v <- c(0,0,0,60,60,60,100)
set_label_game(v)

}
